Support setting commit status
Support commit status updates.  Commit status can be set by a POST to
the appropriate commit URL.  The status can be updated by a subsequent
POST to the same URL with the same `name` and `ref`, but different
values for `state`, `description`, etc.

Note: Listing the commit statuses is not yet supported.  This is done
through a different path on the server, under the `repository` path.

Example of use from the CLI:

    # add a build status to a commit
    gitlab project-commit-status create --project-id 2 \
        --commit-id a43290c --state success --name ci/jenkins \
        --target-url http://server/build/123 \
        --description "Jenkins build succeeded"