Remove 'next_url' from kwargs before passing it to the cls constructor.
The 'next_url' argument causes problems in the _construct_url method if it
doesn't belong there. E.g. if you list all projects, change an attribute
of a project and then try to save it, the _construct_url will use the
'next_url' from the list method and the save will fail.