Rework the manager and object classes
Add new RESTObject and RESTManager base class, linked to a bunch of
Mixin class to implement the actual CRUD methods.

Object are generated by the managers, and special cases are handled in
the derivated classes.

Both ways (old and new) can be used together, migrate only a few v4
objects to the new method as a POC.

TODO: handle managers on generated objects (have to deal with attributes
in the URLs).