Raise an exception on https redirects for PUT/POST
POST and PUT requests are modified by clients when redirections happen.
A common problem with python-gitlab is a misconfiguration of the server
URL: the http to https redirection breaks some requests.

With this change python-gitlab should detect problematic redirections,
and raise a proper exception instead of failing with a cryptic error.

Closes #565