fix: only append kwargs as query parameters
Some arguments to `http_request` were being read
from kwargs, but kwargs is where this function
creates query parameters from, by default. In
the absence of a `query_parameters` param, the
function would construct URLs with query
parameters such as `retry_transient_errors=True`
despite those parameters having no meaning to
the API to which the request was sent.

This change names those arguments that are
specific to `http_request` so that they do not
end up as query parameters read from kwargs.