feat: remove support for Python 3.6, require 3.7 or higher
Python 3.6 is End-of-Life (EOL) as of 2021-12 as stated in
https://www.python.org/dev/peps/pep-0494/

By dropping support for Python 3.6 and requiring Python 3.7 or higher
it allows python-gitlab to take advantage of new features in Python
3.7, which are documented at:
https://docs.python.org/3/whatsnew/3.7.html

Some of these new features that may be useful to python-gitlab are:
  * PEP 563, postponed evaluation of type annotations.
  * dataclasses: PEP 557 – Data Classes
  * importlib.resources
  * PEP 562, customization of access to module attributes.
  * PEP 560, core support for typing module and generic types.
  * PEP 565, improved DeprecationWarning handling

BREAKING CHANGE: As of python-gitlab 3.0.0, Python 3.6 is no longer
supported. Python 3.7 or higher is required.