chore: generate artifacts for the docs build in the CI
When building the docs store the created documentation as an artifact
so that it can be viewed.

This will create a html-docs.zip file which can be downloaded
containing the contents of the `build/sphinx/html/` directory. It can
be downloaded, extracted, and then viewed. This can be useful in
reviewing changes to the documentation.

See https://github.com/actions/upload-artifact for more information on
how this works.