fix: use url-encoded ID in all paths ALTERNATE METHOD
An alternative to https://github.com/python-gitlab/python-gitlab/pull/1819

Make sure all usage of the ID in the URL path is encoded. Normally it
isn't an issue as most IDs are integers or strings which don't contain
a slash ('/'). But when the ID is a string with a slash character it
will break things.

Add a test case that shows this fixes wikis issue with subpages which
use the slash character.

Closes: #1079