chore: make `Gitlab.http_request()` a private method
Convert `Gitlab.http_request()` to `Gitlab._http_request()` to signify
it is a private/protected method so that users of the library know
they should not use the method and we make no API stability promises
for using it.

Add a `Gitlab.http_request()` method which will issue a Deprecation
warning when called. It will pass the call onto
`Gitlab._http_request()`

Also, in the interest of improving code read-ability, require keyword
arg usage when calling `Gitlab._http_request()`