chore(ci): specify name of "stale" label
Saw the following error in the log:
  [#2618] Removing the label "Stale" from this issue...
  ##[error][#2618] Error when removing the label: "Label does not exist"

My theory is that the case doesn't match ("Stale" != "stale") and that
is why it failed.  Our label is "stale" so update this to match.
Thought of changing the label name on GitHub but then would also
require a change here to the "any-of-labels". So it seemed simpler to
just change it here.

It is confusing though that it detected the label "stale", but then
couldn't delete it.