chore: add Python 3.13 as supported & add Python 3.14 CI
Mark that Python 3.13 is supported.

Add a job to test the development versions of Python 3.14.

Add py314 to the tox environment list

Even though there isn't a Python 3.14 at this time, this is done for
the future.  tox is already configured to just warn about missing
Python versions, but not fail if they don't exist.