feat(api): Add argument that appends extra HTTP headers to a request
Currently the only way to manipulate the headers for a request
is to use `Gitlab.headers` attribute. However, this makes it
very concurrently unsafe because the `Gitlab` object can be shared
between multiple requests at the same time.

Instead add a new keyword argument `extra_headers` which will update
the headers dictionary with new values just before the request is sent.

For example, this can be used to download a part of a artifacts file
using the `Range` header: https://developer.mozilla.org/en-US/docs/Web/HTTP/Range_requests

Signed-off-by: Igor Ponomarev <igor.ponomarev@collabora.com>