feat!: drop Python 3.9 support and add Python 3.14
Python 3.9 is End-of-Life (EOL) as of 2025-10 as stated in
https://devguide.python.org/versions/ and
https://peps.python.org/pep-0596/#lifespan

By dropping support for Python 3.9 and requiring Python 3.10 or higher
it allows python-gitlab to take advantage of new features in Python
3.10, which are documented at:
https://docs.python.org/3/whatsnew/3.10.html

Also mark Python 3.14 as officially supported as it has been released.

Closes: https://github.com/python-gitlab/python-gitlab/issues/3285

BREAKING CHANGE: As of python-gitlab 7.0.0, Python 3.9 is no longer
supported. Python 3.10 or higher is required.